+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 2.0.2
-* Add `DynamicTransducer`, and make it default
+* Add `DynamicTransducer` for C-API
+* Remove legacy C-API library
+* Change `AmplitudeConstraint` api in Holo Gain in C-API
 * Fix `wavelength` and `wavenumber` of `NormalTransducer`
 
 # 2.0.1